backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-multipart
pdfplumber
python-docx